Move psychometric-calc module to general/utilities
