+++ src/app/calculator/utilities/fan-psychometric/fan-psychometric-form/fan-psychometric-form.component.spec.ts
