+++ src/app/calculator/utilities/fan-psychometric/fan-psychometric-help/fan-psychometric-help.component.spec.ts
