+++ src/app/calculator/utilities/fan-psychometric/fan-psychometric-results/fan-psychometric-results.component.css
