+++ src/app/calculator/utilities/fan-psychometric/fan-psychometric-table/fan-psychometric-table.component.spec.ts
