+++ src/app/calculator/utilities/fan-psychometric/fan-psychometric.component.html
