+++ src/app/calculator/utilities/fan-psychometric/fan-psychometric.module.ts
