+++ src/app/calculator/utilities/fan-psychometric/fan-psychometric.service.spec.ts
